refactor: Introduce core package
