+++ src/opossum_lib/core/opossum_model.py
