+++ tests/core/test_opossum_model.py
