+++ tests/opossum_model_generators/external_attribution_source_provider.py
@@ -7,7 +7,7 @@ from faker.providers import BaseProvider
 from faker.providers.misc import Provider as MiscProvider
 from faker.providers.person import Provider as PersonProvider
 
-from opossum_lib.opossum_model import ExternalAttributionSource
+from opossum_lib.core.opossum_model import ExternalAttributionSource
 from tests.util.generator_helpers import entry_or_none
